Experimental patches
The ImageMagick configure script runs a number of compiler tests to check
for the existence of libraries such as libxext and libxt (-lXext -lXt).

When the configure script is checking for XShmAttach in -lXext it does so
by testing that -lXext is present but without testing if
X11/extensions/XShm.h can be included.

This check enables MAGICKCORE_HAVE_SHARED_MEMORY
This causes MagickCore/xwindow.c to include X11/extensions/XShm.h
This causes the `make install` step to fail

As an experiment, the configure script is patched to test for the existence
of X11/extensions/XShm.h (see the bottom of this file).

In addition, the configure script is patched to align with a vanilla build
of ImageMagick.

These patches allows the project to build but the `display wizard`
segfaults.